+++ src/gmt_stat.c
@@ -623,7 +623,7 @@ double GMT_in (struct GMT_CTRL *GMT, unsigned int n, double x)
 	tox = 2.0 / fabs (x);
 	bip = res = 0.0;
 	bi = 1.0;
-	m = 2 * (n + lrint (sqrt ((double)(IACC * n))));
+	m = 2 * (n + (unsigned int)lrint (sqrt ((double)(IACC * n))));
 	for (j = m; j >= 1; j--) {
 		bim = bip + ((double)j) * tox * bi;
 		bip = bi;
@@ -1330,7 +1330,7 @@ double GMT_tcrit (struct GMT_CTRL *GMT, double alpha, double nu)
 	}
 	t_low = GMT_zcrit (GMT, alpha);
 	t_high = 5.0;
-	NU = lrint(nu);
+	NU = (int)lrint(nu);
 	GMT_student_t_a (GMT, t_high, NU, &p_high);
 	while (p_high < p) {	/* Must pick higher starting point */
 		t_high *= 2.0;
@@ -1409,8 +1409,8 @@ double GMT_Fcrit (struct GMT_CTRL *GMT, double alpha, double nu1, double nu2)
 	p = 1.0 - alpha;
 	F_low = 0.0;
 	gmt_F_to_ch1_ch2 (GMT, F_high, nu1, nu2, &chisq1, &chisq2);
-	NU1 = lrint (nu1);
-	NU2 = lrint (nu2);
+	NU1 = (int)lrint (nu1);
+	NU2 = (int)lrint (nu2);
 	GMT_f_q (GMT, chisq1, NU1, chisq2, NU2, &p_high);
 	while (p_high > p) {	/* Must pick higher starting point */
 		F_high *= 2.0;
